+++ frontend/src/screens/HomeScreen.js
@@ -1,7 +1,9 @@
-import React, {useState} from 'react';
+import React from 'react';
 
 import "../css/main.css";
 import "../css/buttons.css";
+import "../css/whyBlocks.css";
+import "../css/purpleBlock.css";
 
 import Prices from "../components/Prices";
 import FormUser from "../components/FormUser";
@@ -13,7 +15,7 @@ import Brands from "../components/Brands";
 import ListOfRec from "../components/ListOfRec";
 
 
-import dataForAccordion from "../components/dataForAccordion"
+import dataForAccordion from "../components/dataForAccordion";
 
 
 const HomeScreen = () => {
@@ -23,7 +25,7 @@ const HomeScreen = () => {
             <div id="first-block">
                 <div className="first-block-text">
                     <h2> Негативные отзывы теперь в прошлом!</h2>
-                    <h1><b> <span>RecTOP</span> - инновационный сервис</b> <i>взаимодействия с клиентами.</i></h1>
+                    <h1><b> <span className="gradient-text">RecTOP</span> - инновационный сервис</b> <span className="cursive-text">взаимодействия с клиентами.</span></h1>
                     <br/><br/>
                     <div className="small-text">
                         Отзывы собираются по QR-коду, когда клиент посещает заведение или сайт.
@@ -34,10 +36,12 @@ const HomeScreen = () => {
                 <img
                     src="/images/main/mainCard1Img.png"
                     className="first-image"
+                    alt="first"
                 />
                 <img
                     src="/images/main/mainCard1ImgMobile.png"
                     className="first-img-for-mobile"
+                    alt="first for mobile"
                 />
             </div>
             {/*
@@ -49,7 +53,7 @@ const HomeScreen = () => {
     */}
             <div id="about">
                 <div className="about-comp-text">
-                    <h1><b>О компании <span>Rectop</span></b></h1>
+                    <h1><b>О компании <span className="gradient-text">Rectop</span></b></h1>
 
                     <h2>Качественно повышаем вашу репутацию</h2>
                     <div className="small-text">
@@ -71,22 +75,34 @@ const HomeScreen = () => {
                 <h1><b>Почему RecTop</b></h1>
                 <div className="why-blocks">
                     <div className="why-block">
-                        <img src="/images/pic1.png"/>
+                        <img
+                            src="/images/pic1.png"
+                            alt="mobile"
+                        />
                         <h2><b>Увеличим</b></h2>
                         <p><br/> положительные отзывы реальных клиентов на 50%</p>
                     </div>
                     <div className="why-block">
-                        <img src="/images/pic2.png"/>
+                        <img
+                            src="/images/pic2.png"
+                            alt="mobile"
+                        />
                         <h2><b>Освободим</b></h2>
                         <p><br/> 90% вашего времени на сбор клиентских отзывов</p>
                     </div>
                     <div className="why-block">
-                        <img src="/images/pic3.png"/>
+                        <img
+                            src="/images/pic3.png"
+                            alt="mobile"
+                        />
                         <h2><b>Мотивируем</b></h2>
                         <p><br/> клиентов на повторные продажи: кэшбэк за отзывы</p>
                     </div>
                     <div className="why-block">
-                        <img src="/images/pic4.png"/>
+                        <img
+                            src="/images/pic4.png"
+                            alt="mobile"
+                        />
                         <h2><b>Регулируем</b></h2>
                         <p><br/> конфликт вовремя, негативные отзывы получаете только Вы</p>
                     </div>
@@ -94,7 +110,7 @@ const HomeScreen = () => {
             </div>
             <div className="get-reports">
                 <h1>
-                    <b>Получайте <span>положительные</span> отзывы</b>
+                    <b>Получайте <span className="gradient-text">положительные</span> отзывы</b>
                 </h1>
                 <div className="get-reports-text">
                     Наша команда ежедневно просматривает и анализирует крупнейшие популярные сервисы по сбору отзывов
@@ -108,6 +124,7 @@ const HomeScreen = () => {
                 <img
                     src="/images/mobileforpurpleblock.png"
                     className="img-in-purple-block"
+                    alt="mobile"
                 />
                 <div className="wrapper-for-purple-block">
                     <div className="purple-block">
@@ -121,13 +138,14 @@ const HomeScreen = () => {
                         <img
                             src="/images/mobileforpurpleblockt.png"
                             className="img-in-purple-block-mobile"
+                            alt="mobile half"
                         />
                     </div>
                 </div>
             </div>
             <div id="work">
-                <h1><b>Как мы <span>работаем?</span></b></h1>
-                <h2>Сервис “Генератор отзывов”</h2>
+                <h1><b>Как мы <span className="gradient-text">работаем?</span></b></h1>
+                <h3>Сервис “Генератор отзывов”</h3>
                 <div className="small-text">
                     Уникальный генератор отзывов о компании. Эффективно и просто: клиент сканирует QR-код, оставляет
                     отзыв, вы получаете уведомления в Telegram.
@@ -135,11 +153,13 @@ const HomeScreen = () => {
                 <div className="block-with-pic-one">
                     <img
                         src="/images/main/mainCard51Img.png"
+                        className ="pic-in-block-one"
+                        alt="mobile"
                     />
                     <div className="block-with-pic-one-text">
-                        <h1><b>Акцент на <span>положительных отзывах</span></b></h1>
-                        <h2>Клиенту все понравилось?</h2>
-                        <div>
+                        <h2><b>Акцент на <span className="gradient-text">положительных отзывах</span></b></h2>
+                        <h3>Клиенту все понравилось?</h3>
+                        <div className="small-text">
                             <br/>Довольные клиенты смогут оставлять отзывы в личной карточке компании – их увидят тысячи
                             людей!
                         </div>
@@ -147,9 +167,14 @@ const HomeScreen = () => {
                     </div>
                 </div>
                 <div className="block-with-pic-two">
+                    <img
+                        src="/images/main/mainCard52Img.png"
+                        className="block-two-pic-mobile"
+                        alt="mobile"
+                    />
                     <div className="block-with-pic-two-text">
-                        <h1><b>Негативные отзывы - <span>в службу поддержки</span></b></h1>
-                        <h2>Клиент недоволен?</h2>
+                        <h2><b>Негативные отзывы - <span className="gradient-text">в службу поддержки</span></b></h2>
+                        <h3>Клиент недоволен?</h3>
                         <div>
                             <ul>
                                 <li>Устраняйте конфликтные ситуации в чате службы поддержки в Telegram.</li>
@@ -160,11 +185,13 @@ const HomeScreen = () => {
                     </div>
                     <img
                         src="/images/main/mainCard52Img.png"
+                        className="block-two-pic"
+                        alt="mobile"
                     />
                 </div>
             </div>
             <div className="recommendations">
-                <h1><b><span>Как ликвидировать</span> негативные отзывы?</b></h1>
+                <h1><b><span className="gradient-text">Как ликвидировать</span> негативные отзывы?</b></h1>
                 <h2>Один негативный отзыв - это 100 положительных, и если вы хотите его убрать, то лучше ответить на
                     него в момент возникновения</h2>
                 <ListOfRec/>
@@ -181,7 +208,7 @@ const HomeScreen = () => {
 
             <div id="contacts">
                 <div className="text-contacts">
-                    <h1><b>Запишитесь на <span>бесплатный звонок</span></b></h1>
+                    <h1><b>Запишитесь на <span className="gradient-text">бесплатный звонок</span></b></h1>
                     <p>Наш менеджер перезвонит вам через 10 минут, в течение которых вы сможете:</p>
                     <ul>
                         <li>Узнать все отзывы о своей компании и как их отработать.</li>
@@ -189,8 +216,9 @@ const HomeScreen = () => {
                         <li> Проанализировать общий рейтинг компании и каждой точки продаж в отдельности.</li>
                         <li>Узнать откуда берутся негативные отзывы и как удалить негативный отзыв без бюджета.</li>
                     </ul>
-                    <p>Rectop - Ваша репутация!</p>
+                    <p><b>Rectop - Ваша репутация!</b></p>
                 </div>
+                <div className="text-contacts-mobile"><b>Подарим доступ к сервису “Электронная книга отзывов и предложений” навсегда!</b></div>
                 <FormUser/>
             </div>
             <Refprogram/>